Improved setup and added new targets.
Improved setup, now it handles preparing the servers for use with
ns-scripts-manager. It automatically handles creating symlinks to either
the RELEASE or TESTING directory, based on your configuration. This uses
the servers.txt file as input to configure the servers.
Also added new targets: relink and unlink. Relink rereads the
servers.txt and updates symlinks if necessary, this is useful when you
change server's type and want ns-scripts-manager to update the symlink
of that server.
Unlink deletes every symlink and restores each servers' scripts
directory from the backup (which is created at setup).